新增接口统计控制层代码

+++ msip-bhu-business-impl/msip_bhu_rpc_api/src/main/java/com/bhu/vas/api/rpc/payment/dto/PaymentRecordDTO.java
@@ -7,15 +7,15 @@ package com.bhu.vas.api.rpc.payment.dto;
  *
  */
 public class PaymentRecordDTO {
-	private Float amount;
+	private int amount;
 	private int count;
 	private String info;
 
-	public Float getAmount() {
+	public int getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Float amount) {
+	public void setAmount(int amount) {
 		this.amount = amount;
 	}
